requirements.txt
SpeechRecognition==3.10.0
PyAudio==0.2.14
psutil==5.9.5
pyautogui==0.9.54
google-generativeai==0.5.0
python-dotenv==0.21.1
openai==1.12.0
flet==0.22.0
requests==2.31.0
geopy==2.4.1
